Move map generation to seperate class Move pathfining to seperate class Remove concept of faction leader (for now) Add options to have multiple actors per faction be controlled properly Rework Sighted to cache values of visible tiles Several cleanups and optimizations
