+++ Assets/Actor/ActorDisplay.cs
@@ -11,8 +11,8 @@ public class ActorDisplay : MonoBehaviour, IPointerDownHandler
 
     public void OnPointerDown(PointerEventData data)
     {
+        SystemController.Instance.SetSelectedActor(Actor);
         ActorController.Instance.ShowActorPanel(Actor);
-
         CameraController.Instance.MoveToViewCell(Actor.Location);
     }
