+++ Assets/TopPanelController.cs
@@ -23,16 +23,6 @@ public class TopPanelController : MonoBehaviour
 
     private void Update()
     {
-        if (_player == null)
-        {
-            _player = ActorController.Instance.PlayerFaction.Leader.GetTrait<Player>();
-        }
-
-        if (_player != null)
-        {
-            transform.Find("Available Actions").GetComponent<Text>().text =
-                "Available Actions: " + ActorController.Instance.PlayerFaction.Leader.ActionsAvailable;
-            transform.Find("Gold").GetComponent<Text>().text = "Gold: " + Gold;
-        }
+     
     }
 }
